requirements.txt
streamlit
pandas
numpy
pydeck
altair
requests
python-dotenv
pyarrow        # parquet 저장/로드
